refactor: moved some stuff into components
